feat(space): lil fixes
