+++ packages/space/tsconfig.json
@@ -3,10 +3,6 @@
     "extends": "../../tsconfig.json",
     "compilerOptions": {
         "outDir": "dist",
-        "rootDir": "src",
-        "baseUrl": ".",
-        "paths": {
-            "@alfalab/core-components-*": ["../*/src"]
-        }
+        "rootDirs": ["src"]
     }
 }
